+++ graphql/server-test/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.1](https://github.com/constructive-io/constructive/compare/graphql-server-test@2.3.0...graphql-server-test@2.3.1) (2026-03-03)
+
+**Note:** Version bump only for package graphql-server-test
+
 # [2.3.0](https://github.com/constructive-io/constructive/compare/graphql-server-test@2.2.5...graphql-server-test@2.3.0) (2026-03-01)
 
 **Note:** Version bump only for package graphql-server-test